+++ pages/student.php
@@ -180,13 +180,13 @@
 					let response = JSON.parse(e);
 					if(response.status == "success"){
 						alert(response.message);
+						$('#table').dataTable().fnClearTable();
+						$('#table').dataTable().fnDestroy();
+						displayData(); 
+						$(".form-control").val("");
 					} else {
 						alert(response.message);
 					}
-					$('#table').dataTable().fnClearTable();
-					$('#table').dataTable().fnDestroy();
-					displayData(); 
-					$(".form-control").val("");
 				}
 			});
 		})